Add /*#__PURE__*/ annotations to hint Terser for tree-shaking
Webpack is more conservative in terms of dead code elimination
compared to Rollup, which causes it to bundle all the icons even
when they are not used. This patch annotate the `genIcon` and
`Object.keys` function as pure to give Terser a hint that it is
in fact side effect free.

Closes #1.